refactor(core): split runtime messages into package
